requirements.txt
customtkinter
tkinterdnd2
pymupdf
pillow
python-pptx
python-docx
pyinstaller